requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
requests==2.28.1
circuitbreaker==1.4.0
python-dotenv
pyjwt
git+https://github.com/ozgur/python-firebase
google-api-python-client
google-auth-httplib2
google-auth-oauthlib